About
Prestige Coding and Technology Limited is a born out of passion for technology Computer Coding/Programming Institute and Software development company. It was established to provide the services of building efficient and effective software and applications for businesses, companies, and individuals. Its academic  arm provides  coding and programming education for beginners, professionals and organizations.